# api.yaml
openapi: 3.0.0
info:
  title: Kanban API
  version: 0.2.0
  description: |
    상담 칸반 보드 API. 멀티테넌트 환경 (병원별 로그인 기반).

    - JWT Bearer 인증 (Authorization: Bearer <token>) 사용.
    - `/api/auth/login` 만 public, 그 외 모든 엔드포인트는 인증 필요.
    - 관리자(`is_admin=true`) 계정만 `/api/admin/hospitals` 접근 가능.
    - 날짜별로 컬럼 구조(row1/row2)와 카드를 관리합니다.
    - 템플릿을 통해 컬럼 구조를 다른 날짜에 적용할 수 있습니다.
    - 동일 차트번호 카드 간 형제 동기화(propagation) 기능을 제공합니다.
servers:
  - url: /
    description: Default server
...
security:
  - BearerAuth: []
paths:
  /api/auth/login:
    post:
      tags: [auth]
      summary: 병원 계정 로그인
      description: |
        이메일/비밀번호로 로그인하여 JWT 액세스 토큰을 발급받습니다.
        반환된 access_token 을 이후 요청의 `Authorization: Bearer <token>` 헤더에 사용합니다.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              email: hospital@example.com
              password: secret1234
      responses:
        '200':
          description: 로그인 성공 (토큰 + 병원 공개 정보)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/auth/me:
    get:
      tags: [auth]
      summary: 현재 로그인 병원 정보 조회
      description: |
        Authorization 헤더의 토큰으로부터 현재 로그인한 병원의 공개 정보를 반환합니다.
        password_hash 등 민감 필드는 절대 포함되지 않습니다.
      responses:
        '200':
          description: 현재 로그인 병원 공개 정보
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HospitalPublic'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/boards/{date}:
    get:
      tags: [boards]
      summary: 일자별 보드 스냅샷 조회
      description: |
        지정한 날짜의 컬럼 구조(row1/row2)와 카드 전체를 한 번에 조회합니다.
        형제 카드 propagation(동일 차트번호 동기화)이 적용된 상태로 반환됩니다.
      parameters:
        - $ref: '#/components/parameters/DatePath'
      responses:
        '200':
          description: 보드 스냅샷
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BoardSnapshot'
              example:
                date: '2026-04-27'
                columns:
                  date: '2026-04-27'
                  row1:
                    - { id: r1_구환, date: '2026-04-27', label: 구환상담, position: 0, built_in: true }
                    - { id: r1_신환, date: '2026-04-27', label: 신환상담, position: 1, built_in: true }
                  row2:
                    - { id: r2_구환, date: '2026-04-27', row1_id: r1_구환, label: 구환, position: 0, built_in: true }
                    - { id: r2_신환, date: '2026-04-27', row1_id: r1_신환, label: 신환, position: 0, built_in: true }
                cards:
                  - id: card_001
                    date: '2026-04-27'
                    row1_id: r1_신환
                    row2_id: r2_신환
                    time: '10:00'
                    name: 홍길동
                    chart: '12345'
                    counselor: 김수지
                    book_time: '10:00'
                    consult_time: '15'
                    memo: ''
                    color: ''
                    created_at: '2026-04-27T09:00:00Z'
                    updated_at: '2026-04-27T09:00:00Z'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/boards/{date}/apply-template:
    post:
      tags: [boards]
      summary: 템플릿을 보드에 적용
      description: |
        지정한 템플릿의 컬럼 구조를 해당 날짜에 적용합니다.
        기존 컬럼은 교체되며, 적용 후의 보드 스냅샷이 반환됩니다.
      parameters:
        - $ref: '#/components/parameters/DatePath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApplyTemplateRequest'
            example:
              template_id: tpl_default
      responses:
        '200':
          description: 템플릿이 적용된 보드 스냅샷
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BoardSnapshot'
        '404':
          $ref: '#/components/responses/NotFound'
        '400':
          $ref: '#/components/responses/BadRequest'
...
  /api/cards:
    get:
      tags: [cards]
      summary: 일자별 카드 목록 조회
      description: 지정한 날짜의 카드 전체를 반환합니다.
      parameters:
        - name: date
          in: query
          required: true
          schema: { type: string, format: date }
          description: 조회 일자 (YYYY-MM-DD)
      responses:
        '200':
          description: 카드 목록
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Card'
    post:
      tags: [cards]
      summary: 카드 생성
      description: |
        새로운 카드를 생성합니다.
        single_card_r1 에 속한 row1 의 동일 시간 슬롯은 카드 1개만 허용됩니다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CardCreate'
            example:
              date: '2026-04-27'
              row1_id: r1_신환
              row2_id: r2_신환
              time: '10:00'
              name: 홍길동
              chart: '12345'
              counselor: 김수지
              book_time: '10:00'
              consult_time: '15'
              memo: ''
              color: ''
      responses:
        '201':
          description: 생성된 카드
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Card'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/cards/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string }
        description: 카드 ID
    patch:
      tags: [cards]
      summary: 카드 수정
      description: |
        카드 필드를 수정합니다.
        sync_siblings=true 이면 동일 차트번호의 형제 카드들도 동일 필드로 동기화됩니다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CardUpdate'
            example:
              name: 홍길동(수정)
              chart: '12345'
              counselor: 라근주
              memo: 메모 변경
              sync_siblings: true
      responses:
        '200':
          description: 수정된 카드
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Card'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [cards]
      summary: 카드 삭제
      responses:
        '204':
          description: 삭제 성공
        '404':
          $ref: '#/components/responses/NotFound'
  /api/cards/{id}/move:
    patch:
      tags: [cards]
      summary: 카드 이동
      description: |
        카드의 시간/소속 컬럼(row1/row2)을 변경합니다.
        single_card_r1 컬럼 내 충돌 시 409.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
          description: 카드 ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CardMove'
            example:
              time: '11:30'
              row1_id: r1_임플
              row2_id: r2_임플_상
      responses:
        '200':
          description: 이동된 카드
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Card'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/cards/by-chart:
    get:
      tags: [cards]
      summary: 차트번호로 형제 카드 조회
      description: |
        동일 차트번호를 가진 카드들을 조회합니다 (formula propagation 용).
        date 가 주어지면 해당 일자로 한정합니다.
      parameters:
        - name: chart
          in: query
          required: true
          schema: { type: string }
          description: 차트번호
        - name: date
          in: query
          required: false
          schema: { type: string, format: date }
          description: 조회 일자 (옵션)
      responses:
        '200':
          description: 동일 차트번호 카드 목록
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Card'
  /api/stats/implant:
    get:
      tags: [stats]
      summary: 임플란트 통계 조회
      description: 해당 일자의 임플란트(row1=r1_임플) 카드 수를 row2 별로 집계합니다.
      parameters:
        - name: date
          in: query
          required: true
          schema: { type: string, format: date }
          description: 조회 일자
      responses:
        '200':
          description: 임플란트 통계
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImplantStats'
              example:
                date: '2026-04-27'
                total: 5
                by_leaf:
                  r2_임플_상: { label: 상, count: 2 }
                  r2_임플_중: { label: 중, count: 2 }
                  r2_임플_하: { label: 하, count: 1 }
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        로그인 후 발급받은 JWT 토큰을 `Authorization: Bearer <token>` 헤더에 담아 전송합니다.
  parameters:
    DatePath:
      name: date
      in: path
      required: true
      schema:
        type: string
        format: date
      description: 일자 (YYYY-MM-DD)
      example: '2026-04-27'
  responses:
    BadRequest:
      description: 잘못된 요청
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: 인증 실패 (토큰 누락/만료/유효하지 않음, 또는 로그인 자격 증명 불일치)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    NotFound:
      description: 리소스를 찾을 수 없음
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: 제약 충돌 (built_in 삭제, 단일카드 슬롯 충돌, 기본 템플릿 삭제, 이메일 중복 등)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: 에러 코드
          example: NOT_FOUND
        message:
          type: string
          description: 에러 메시지 (사용자 노출용)
          example: 리소스를 찾을 수 없습니다.
        detail:
          type: string
          nullable: true
          description: 상세 디버그 정보 (옵션)
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          description: 병원 계정 이메일
          example: hospital@example.com
        password:
          type: string
          minLength: 8
          description: 병원 계정 비밀번호 (평문, HTTPS 전송 가정)
          example: secret1234
...
    HospitalPublic:
      type: object
      description: |
        병원 공개 정보 (로그인 응답 / 현재 세션 정보용).
        password_hash, is_active, created_at, updated_at 등은 노출하지 않는다.
      required: [id, name, email, is_admin]
      properties:
        id:
          type: string
          example: hosp_001
        name:
          type: string
          example: 강남치과
        email:
          type: string
          format: email
          example: gangnam@example.com
        is_admin:
          type: boolean
          example: false
...
    AuthResponse:
      type: object
      description: 로그인 성공 시 반환되는 토큰 + 병원 공개 정보.
      required: [access_token, hospital]
      properties:
        access_token:
          type: string
          description: JWT Bearer 토큰
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        hospital:
          $ref: '#/components/schemas/HospitalPublic'
    ColumnRow1:
      type: object
      required: [id, date, label, position, built_in]
      properties:
        id:
          type: string
          description: row1 컬럼 ID
          example: r1_신환
        date:
          type: string
          format: date
          description: 일자
          example: '2026-04-27'
        label:
          type: string
          description: 컬럼 라벨
          example: 신환상담
        position:
          type: integer
          description: 표시 순서 (0부터 시작)
          example: 1
        built_in:
          type: boolean
          description: 시스템 기본 컬럼 여부 (true 면 삭제 불가)
          example: true
...
    ColumnRow2:
      type: object
      required: [id, date, row1_id, label, position, built_in]
      properties:
        id:
          type: string
          example: r2_임플_상
        date:
          type: string
          format: date
        row1_id:
          type: string
          description: 부모 row1 컬럼 ID
          example: r1_임플
        label:
          type: string
          example: 상
        position:
          type: integer
          example: 0
        built_in:
          type: boolean
          example: true
...
    ColumnsBundle:
      type: object
      required: [date, row1, row2]
      properties:
        date:
          type: string
          format: date
        row1:
          type: array
          items:
            $ref: '#/components/schemas/ColumnRow1'
        row2:
          type: array
          items:
            $ref: '#/components/schemas/ColumnRow2'
...
    Card:
      type: object
      required:
        [
          id,
          date,
          row1_id,
          row2_id,
          time,
          name,
          chart,
          counselor,
          book_time,
          consult_time,
          memo,
          color,
          created_at,
          updated_at,
        ]
      properties:
        id:
          type: string
          example: card_001
        date:
          type: string
          format: date
        row1_id:
          type: string
          example: r1_신환
        row2_id:
          type: string
          example: r2_신환
        time:
          type: string
          description: 슬롯 시간 (HH:MM)
          example: '10:00'
        name:
          type: string
          description: 환자 이름
          example: 홍길동
        chart:
          type: string
          description: 차트번호 (형제 동기화 키)
          example: '12345'
        counselor:
          type: string
          description: 상담사
          example: 김수지
        book_time:
          type: string
          description: 예약 시간 (HH:MM 또는 빈 문자열)
          example: '10:00'
        consult_time:
          type: string
          description: 상담 소요 시간 (분 단위 문자열)
          example: '15'
        memo:
          type: string
          description: 메모
        color:
          type: string
          description: 카드 배경 색상 (HEX 또는 빈 문자열)
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CardCreate:
      type: object
      required: [date, row1_id, row2_id, time]
      properties:
        date:
          type: string
          format: date
        row1_id:
          type: string
        row2_id:
          type: string
        time:
          type: string
          description: 슬롯 시간 (HH:MM)
        name:
          type: string
          nullable: true
        chart:
          type: string
          nullable: true
        counselor:
          type: string
          nullable: true
        book_time:
          type: string
          nullable: true
        consult_time:
          type: string
          nullable: true
        memo:
          type: string
          nullable: true
        color:
          type: string
          nullable: true
    CardUpdate:
      type: object
      description: |
        카드 필드 부분 수정. sync_siblings=true 면 동일 chart 의 형제 카드도 동기화됩니다.
      properties:
        name:
          type: string
          nullable: true
        chart:
          type: string
          nullable: true
        counselor:
          type: string
          nullable: true
        book_time:
          type: string
          nullable: true
        consult_time:
          type: string
          nullable: true
        memo:
          type: string
          nullable: true
        color:
          type: string
          nullable: true
        sync_siblings:
          type: boolean
          nullable: true
          description: 동일 차트번호의 형제 카드도 동일 필드로 동기화할지 여부
          example: true
    CardMove:
      type: object
      description: 카드 이동 (시간/소속 컬럼 변경)
      properties:
        time:
          type: string
          nullable: true
          description: 새 슬롯 시간 (HH:MM)
        row1_id:
          type: string
          nullable: true
        row2_id:
          type: string
          nullable: true
...
    BoardSnapshot:
      type: object
      required: [date, columns, cards]
      properties:
        date:
          type: string
          format: date
        columns:
          $ref: '#/components/schemas/ColumnsBundle'
        cards:
          type: array
          items:
            $ref: '#/components/schemas/Card'
    ApplyTemplateRequest:
      type: object
      required: [template_id]
      properties:
        template_id:
          type: string
          description: 적용할 템플릿 ID
          example: tpl_default
    ImplantLeafCount:
      type: object
      required: [label, count]
      properties:
        label:
          type: string
          example: 상
        count:
          type: integer
          example: 2
    ImplantStats:
      type: object
      required: [date, total, by_leaf]
      properties:
        date:
          type: string
          format: date
        total:
          type: integer
          description: 임플란트 카드 총 수
          example: 5
        by_leaf:
          type: object
          description: row2 ID 를 키로 한 집계 결과
          additionalProperties:
            $ref: '#/components/schemas/ImplantLeafCount'
...